NED-AGDISTIS 4.4.0
add new cache test
