+++ qanary-component-NED-AGDISTIS/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
     <artifactId>qanary-component-NED-AGDISTIS</artifactId>
     <groupId>eu.wdaqua.qanary.component</groupId>
-    <version>4.3.1</version>
+    <version>4.4.0</version>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
@@ -14,7 +14,8 @@
     </parent>
     <properties>
         <java.version>17</java.version>
-        <qanary.version>[3.7.0,4.0.0)</qanary.version>
+        <qanary.component.version>[3.7.0,4.0.0)</qanary.component.version>
+        <qanary.commons.version>[3.8.0,4.0.0)</qanary.commons.version>
         <docker.image.prefix>qanary</docker.image.prefix>
         <docker.image.name>qanary-component-ned-agdistis</docker.image.name>
         <dockerfile-maven-version>1.4.13</dockerfile-maven-version>
@@ -24,7 +25,14 @@
         <dependency>
             <groupId>eu.wdaqua.qanary</groupId>
             <artifactId>qa.component</artifactId>
-            <version>${qanary.version}</version>
+            <version>${qanary.component.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>eu.wdaqua.qanary</groupId>
+            <artifactId>qa.commons</artifactId>
+            <version>${qanary.commons.version}</version>
+            <type>test-jar</type>
+            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
